scripts: Added tool to merge mrw format xml's
The device generation are involved with different type
of xml's which are generated as a mrw format xml.
Merging similar format xml and processing the xml will be easy
instead of processing each xml individually to generate device
tree. So, This tool is used merge similar mrw format xml which have
header and footer as "attributes" tag.

Signed-off-by: Ramesh Iyyar <rameshi1@in.ibm.com>