data/p10: Added ATTR_SYS_CLOCK_DECONFIG_STATE
ATTR_SYS_CLOCK_DECONFIG_STATE contains the functional state
of the system clocks/oscillators. This is the way that Hostboot
communicates to the BMC which clocks have failed.
Hostboot does not model the clocks as targets, which is why
we need a new attribute.

A_DECONFIG means OSC1 is active, B_DECONFIG means OSC0 is active.
"A" corresponds to the CP_REFCLOCK_SELECT value of OSC0 and
"B" to OSC1.

At the beginning of the IPL, this attribute will be set to
reflect the CP_REFCLOCK_SELECT scratch register, which tells
Hostboot which oscillator to use. If a clock operation fails
and an error log deconfigures an oscillator, the attribute will
be updated to reflect which clocks have failed so far, and this
attribute will be synced to the BMC.

Signed-off-by: Jayanth Othayoth <ojayanth@in.ibm.com>