dt-api : do not log error trace if attribute not found in target
Trying to read ATTR_HWAS_STATE property of all the targets present
in the system using pdbg_target_traverse recursive method noticied
lot of error traces being logged as the ATTR_HWAS_STATE is not
supported in some of the targets.

At present if an attribute is not present in the target the method
returns error FAPI2_RC_INVALID_ATTR_GET, with which the caller
can choose to skip the target or log error trace.

Let the caller decide if an error message needs to be logged
based on the error code rather than the api.

Signed-off-by: Marri Devender Rao <devenrao@in.ibm.com>
Change-Id: Ie18d6432ef58d20618be2bbd8852fc883737a9ea