Fix dimm index numbers in bonnell
Regardless of the DIMM type, whether it be DDR4 or DDR5, the index value
assigned to the DIMMs must correspond to the FAPI_POS.

Also for the Perv targets, the index numbers are the same. So use FAPI_POS as
the index value.

Test results:
Tested and works as expected.